requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
fonttools==4.61.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pdf2image==1.17.0
pillow==12.1.1
pyparsing==3.3.2
pypdf==6.7.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
requests==2.32.5
seaborn==0.13.2
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.6
